cleaning up

+++ Components/recorder.js
@@ -1,8 +1,9 @@
 import React from "react";
 import { RNVoiceRecorder } from "react-native-voice-recorder";
-import { StyleSheet, Text, View, TextInput, TouchableOpacity, Image, ImageBackground, ScrollView, Alert} from "react-native";
+import { StyleSheet, Text, View, TextInput, TouchableOpacity, Image, ImageBackground, ScrollView, Alert } from "react-native";
 import { _onDeleteDictation } from "./Alerts/warnings";
 import { withFirebaseHOC } from '../Firebase';
+import { _handleStoreErr } from "./Alerts/errors";
 
 let recordingPath;
 class Recorder extends React.Component{
@@ -11,6 +12,14 @@ class Recorder extends React.Component{
     message: "",
     successMessage: ""
   };
+
+  componentDidMount = () => {
+    this.setState({message: ""});
+  }
+
+  componentDidUpdate = () => {
+    setTimeout(() => this.setState({message: ""}), 4000);
+  }
 
   _onRecord = async () => {
     if(this.state.dictationName == null || this.state.dictationName == ""){
@@ -31,7 +40,7 @@ class Recorder extends React.Component{
             this.setState({successMessage: "Upload success"});
             console.log(uploadTask);
           }catch(error){
-            console.log(error);
+            this.setState({message: _handleStoreErr(error.code)});
           }
         },
         onCancel: () => {
@@ -55,19 +64,20 @@ class Recorder extends React.Component{
         }
       });
     }
-    else{
-      this.setState({message: "Please record and save the audio first before listening to it"});
-    }
+    else this.setState({message: "Please record and save audio first before listening to it!"});
   }
 
   _onDelete = async () => {
-    response = await _onDeleteDictation();
-    if(response){
-      try{
-        await this.props.firebase.deleteDictation(this.state.dictationName);
-        Alert.alert("Dictation was succesfully deleted.");
-      }catch(error){
-
+    if(recordingPath == null) this.setState({message: "There is nothing to delete!"});
+    else{
+      response = await _onDeleteDictation();
+      if(response){
+        try{
+          await this.props.firebase.deleteDictation(this.state.dictationName);
+          Alert.alert("Dictation was succesfully deleted.");
+        }catch(error){
+          this.setState({message: _handleStoreErr(error.code)});
+        }
       }
     }
   }
@@ -101,7 +111,7 @@ class Recorder extends React.Component{
               <TouchableOpacity onPress={this._onRecord}>
                 <Image style={styles.buttons} source={require("../Assets/RecordButtonIdle.png")}/>
               </TouchableOpacity>
-              <Text style={styles.helperTxt}>Recording Playback</Text>
+              <Text style={styles.helperTxt}>Playback  /  Delete</Text>
             </View>
             <View style={styles.dictationView}>
               <TouchableOpacity onPress={this._onPlayback}>
@@ -159,16 +169,20 @@ const styles = StyleSheet.create({
     fontSize: 16
   },
   warningText: {
-    color: "red",
+    color: "white",
     fontSize: 18,
     padding: 5,
-    textAlign: "center"
+    backgroundColor: "red",
+    alignSelf: "center",
+    borderRadius: 15
   },
   successText: {
-    color: "green",
+    color: "white",
     fontSize: 18,
     padding: 5,
-    textAlign: "center"
+    backgroundColor: "green",
+    alignSelf: "center",
+    borderRadius: 15
   }
 });
